backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==3.2.0
bidict==0.23.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
passlib==1.7.4
paypalrestsdk==1.13.3
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
PyMySQL==1.1.2
pyOpenSSL==25.3.0
python-dotenv==1.2.1
python-engineio==4.12.3
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.14.2
requests==2.32.5
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
wsproto==1.2.0
SQLAlchemy==2.0.44
alembic==1.13.1
alembic==1.13.1
psycopg2-binary==2.9.9
